chore: Refactor datasource import paths and remove unused datasource file; update Cloud Run configuration in Terraform
