+++ src/__tests__/fraudPrevention.test.ts
@@ -2,7 +2,7 @@ import {
   FraudPreventionEntity,
   RiskLevel,
 } from "../entity/fraudPreventionEntity";
-import { AppDataSource } from "../datasource/datasource";
+import { AppDataSource } from "../datasource/data-source";
 import { Repository } from "typeorm";
 import { FraudPreventionController } from "../controllers/fraudPreventionController";
 import { Request, Response } from "express";
